cr feedback and moving tests to file
