+++ tests/unittests/CoreGraphics.drawing/data/reference/TestImage.CGGradient..LinearGradient2.png
