+++ tests/unittests/CoreGraphics.drawing/data/reference/TestImage.CGGradient..LinearGradient3.png
