+++ tests/unittests/CoreGraphics.drawing/data/reference/TestImage.CGGradient..LinearGradientShortBothSides.png
