+++ tests/unittests/CoreGraphics.drawing/data/reference/TestImage.CGGradient.LinearGradient.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d1ebed1412be8f4b436ed215764017185bc8da4d9da79e69bb224f0b2d00b668
+size 248700
