+++ tests/unittests/CoreGraphics.drawing/data/reference/TestImage.CGGradient.LinearGradientInvalidCount.png
