+++ tests/unittests/CoreGraphics.drawing/data/reference/TestImage.CGGradient.LinearGradientShortBothSides_Options_kCGGradientDrawsBeforeStartLocation.png
