+++ tests/unittests/CoreGraphics.drawing/data/reference/TestImage.CGGradient.LinearGradientWithAlpha.png
