+++ tests/unittests/CoreGraphics.drawing/data/reference/TestImage.CGGradient.LinearGradientWithLowOpacityShort.png
